services/users/requirements.txt
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
click==8.1.7
fastapi==0.115.6
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
uuid==1.30
uvicorn==0.34.0
uvloop==0.21.0
validate-docbr==1.10.0
watchfiles==1.0.3
websockets==14.1
